Move non-acquisition code to a `core` module
